restarting docker-compose to just run postgress instance
