+++ database.conf
@@ -1,5 +1,5 @@
 POSTGRES_USER=user1
 POSTGRES_PASSWORD=password
-POSTGRES_HOST=http://127.0.0.1
-POSTGRES_PORT=5432
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5000
 POSTGRES_DB=nqueens
