+++ docker-compose.yml
@@ -2,33 +2,6 @@ version: '3.5'
 # version of docker-compose that we are using
 services:
 # containters we want to run
-  front-end:
-  #displays html to the user
-    container_name: front-end
-    build:
-      context: ./front-end
-    environment:
-      - RESULTS_API_SERVER=http://results-api:5000    
-    ports:
-      - 5002:5000
-    volumes:
-      - ./front-end/src:/app
-
-  results-api:
-  #name of service
-    container_name: results-api
-    # provides an API for CRUD* operations to database
-    build:
-    # Docker parses the docker-compose file and builds our services according to the build instructions.
-      context: ./results-api
-      # where the Dockerfile is located at to build the image
-    env_file: database.conf
-    ports:
-      - 5000:5000
-    volumes:
-      - ./results-api/src:/app  
-      # This mounts the .results-api/src directory on our host computer to the /app directory in the container. 
-
   database:
     container_name: postgres
     # provides an instance of postgress database
@@ -40,9 +13,11 @@ services:
     # Inside the container, this directory is where Postgres stores all the relevant tables and databases.
     # We are defining a volume inside the compose file and mounting it at this mount point.
     ports:
-      - 5001:5432 
+      - 127.0.0.1:5000:5432
     # host machine:dbcontainer ports 
     # map ports from the host machine to the container
+    # incoming requests at docker host port 5002 will be forwarded to the port 5432 of 
+    # the database container, where Postgres server can process it.
 
 volumes:
   # For data persistence purposes, we are storing data in a mouted volume which we have named db-data
