--- front-end/Dockerfile
@@ -1,9 +1,0 @@
-FROM  python:3.6 
-
-WORKDIR /app
-
-COPY ./src .
-RUN pip install -r requirements.txt 
-
-EXPOSE 5000
-CMD python server.py
