+++ requirements.txt
@@ -1,3 +1,3 @@
 Flask==0.12
-psycopg2==2.7.6.1
+psycopg2-binary==2.7.6.1
 flask-sqlalchemy==2.3.0
